requirements.txt
# Flask stack
Flask~=3.0.0
Flask-SQLAlchemy~=3.1.0
Flask-WTF~=1.2.0
Flask-Bcrypt~=1.0.1
Flask-Login~=0.6.3

# PDF processing (PyMuPDF — no PyTorch required)
PyMuPDF>=1.23.0
# pdfplumber: layout-aware PDF text extraction (used as primary extractor in training)
pdfplumber>=0.9.0

# Export formats
openpyxl~=3.1.0
python-docx~=1.1.0

# Utilities
python-dotenv~=1.0.1
Werkzeug~=3.0.0

# -----------------------------------------------------------------------
# Automatic Extraction System — AI/ML/IDP dependencies
# -----------------------------------------------------------------------

# Mindee IDP — structured document extraction (invoice, receipt, ID)
mindee>=4.0.0

# OpenAI GPT — LLM-powered semantic extraction
openai>=1.0.0

# Machine Learning — field classifier & doc classifier
scikit-learn>=1.4.0
numpy>=1.26.0

# OCR — Tesseract wrapper + PDF-to-image conversion
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0